Dumped GFT, made charts pretty (See full msg)
Ok, kinda too big of a commit, but whatever for now.

The charts are now stacked and animated upon initial draw. I’d like to
get them to animate between tiers, but that will require tighter
integration with AngularJS so the charts aren’t “drawn for the first
time” every time.

The other big change (yes, two in a commit, bad Ben) is Google Fusion
Tables, for the time being, has been fired. I realize this is a big
change and is running somewhat counter to what we’ve been doing in
class. HOWEVER, considering the data for this project, using a tools
like GFT is way over-engineering and doesn’t have the features we need.
I spent a fair amount of time battling the SQL-like syntax and
programmatically adding and removing annotations for a pretty graph,
But that was stupid, it took me about a minute to whip together a pivot
table in Google Docs and copy & paste it as a JS array. We have a tiny,
static, dataset. This solution is much easier.